+++ src/Components/Navbar/HomeButton.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Button } from "@mui/material";
 import { useTheme } from "@mui/material/styles";
-import "../css/HomeButton.css";
+import "../../css/HomeButton.css";
 import { useNavigate } from "react-router-dom";
 
 /**
