+++ src/Components/Navbar/LoginButton.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { Button } from "@mui/material";
 import { useTheme } from "@mui/material/styles";
 import { useNavigate } from "react-router-dom";
-import "../css/LoginButton.css";
+import "../../css/LoginButton.css";
 
 /**
  * A button component that navigates the user to the login page.
