+++ src/Components/Navbar/ProfileButton.tsx
@@ -2,9 +2,9 @@ import React, { useState } from "react";
 import { Box, Button } from "@mui/material";
 import { useTheme } from "@mui/material/styles";
 import { useNavigate } from "react-router-dom";
-import "../css/ProfileButton.css";
-import { useUser } from "../hooks/UserContext";
-import { logout } from "../utils/account.utils";
+import "../../css/ProfileButton.css";
+import { useUser } from "../../hooks/UserContext";
+import { logout } from "../../utils/account.utils";
 
 
 const ProfileButton: React.FC = () => {
