+++ src/pages/Login.tsx
@@ -4,7 +4,7 @@ import React, { useState } from "react";
 //using a custom react hook for validating password confirmation
 import usePasswordValidation from "../hooks/usePasswordValidation";
 import { red } from "@mui/material/colors";
-import NavBar from "../Components/NavBar";
+import NavBar from "../Components/Navbar/NavBar";
 import { useTheme } from "@mui/material/styles";
 import authUser from "../utils/loginPage.utils";
 import { useNavigate } from "react-router-dom";
